+++ server/src/index.ts
@@ -11,6 +11,7 @@ import createAuthRouter from "./modules/auth/auth.router";
 import createWatchlistsRouter from "./modules/watchlists/watchlists.router";
 import { startPlayersAutoUpdateScheduler } from "./jobs/playersAutoUpdateScheduler";
 import createShadowTeamsRouter from "./modules/shadow-teams/shadow-teams.router";
+import createRecruitmentRouter from "./modules/recruitment/recruitment.router";
 
 async function startServer() {
     if (process.env.NODE_ENV !== "production") dotenv.config();
@@ -44,6 +45,7 @@ async function startServer() {
     server.use("/", createPlayersRouter(context));
     server.use("/watchlists", createWatchlistsRouter(context));
     server.use("/shadow-teams", createShadowTeamsRouter(context));
+    server.use("/recruitment", createRecruitmentRouter(context));
     server.use("/auth", createAuthRouter(context));
 
     const stopPlayersAutoUpdateScheduler = startPlayersAutoUpdateScheduler(context);